add read yml file and test http method for api
